refactor: introduce botcopier package
