+++ scripts/cluster_regimes.py
@@ -22,11 +22,8 @@ except Exception:  # pragma: no cover - optional
     hdbscan = None
 
 # Reuse data loading and feature extraction from training script
-from train_target_clone import _load_logs, _load_calendar  # type: ignore
-try:
-    from scripts.features import _extract_features  # type: ignore
-except Exception:
-    from features import _extract_features  # type: ignore
+from botcopier.data.loading import _load_logs
+from botcopier.features.engineering import _extract_features
 
 
 def cluster_features(
