+++ tests/test_train_target_clone_multi.py
@@ -3,7 +3,9 @@ from pathlib import Path
 
 import numpy as np
 
-from scripts.train_target_clone import _load_logs, _extract_features, train
+from botcopier.data.loading import _load_logs
+from botcopier.features.engineering import _extract_features
+from botcopier.training.pipeline import train
 
 
 def test_multi_horizon_training(tmp_path: Path) -> None:
