+++ tests/test_train_target_clone_quantile_model.py
@@ -1,7 +1,7 @@
 import json
 from pathlib import Path
 
-from scripts.train_target_clone import train
+from botcopier.training.pipeline import train
 
 
 def test_quantile_model_outputs_quantiles(tmp_path: Path) -> None:
